Trait envoy_sdk::host::stream_info::StreamInfo

pub trait StreamInfo {
    fn stream_property(&self, path: &[&str]) -> Result<Option<ByteString>>;

    fn set_stream_property(&self, path: &[&str], value: &[u8]) -> Result<()>;
}

An interface of the Envoy Stream Info API.

Basic usage of StreamInfo:

use envoy::host::StreamInfo;

let stream_info = StreamInfo::default();

let plugin_name = stream_info.stream_property(&["plugin_name"])?;

stream_info.set_stream_property(&["my_extension", "output"], b"property value")?;

Required methods

fn stream_property(&self, path: &[&str]) -> Result<Option<ByteString>>

Evaluates value of a given property in the enclosing context.

In case HttpFilter, the value will be evaluated in the context of HTTP stream.

In case NetworkFilter, the value will be evaluated in the context of TCP connection.

In case AccessLogger, the value will be evaluated in the context of HTTP stream or TCP connection that is being logged.

Arguments

path - property path as an array of path segments

fn set_stream_property(&self, path: &[&str], value: &[u8]) -> Result<()>

Saves a value in the enclosing context.

The value will be accessible to other filters on that HTTP stream or TCP connection.

Arguments

path - property path as an array of path segments

value - an opaque blob of bytes

Loading content...
Implementations

impl dyn StreamInfo

pub fn default() -> &'static dyn StreamInfo

Returns the default implementation that interacts with Envoy through its ABI.